Use custom date picker adapter
